adas-microservices-latest-1/services/daw-service/requirements.txt
paho-mqtt
numpy
scikit-learn
joblib
flask